Checkpoint: pull out main.py
